fix #34 -- use entry_points to get proper scripts
This is a fix to the shebang issue.  The following scripts are created:

* py
* py2 | py3
* py2.6 | py2.7 | py3.3 | py3.4

The shebang is of the interpreter that was used to install.

py2.5 and earlier, py3.2 and earlier, and py3.5 and later are (currently)
unsupported because (a) I’m lazy and (b) there is no way to automate creation
of completion functions.

Bonus: if someone manages to install it under Windows (which is currently
impossible), they will also get .exe versions of those (that will possibly
collide with py.exe shipped with Python 3.x depending on %PATH% — see #31).

Signed-off-by: Chris “Kwpolska” Warrick <kwpolska@gmail.com>